#431 remove obsolete dependency
